moved item adapter into standalone class
